refactor: correct separation of urls
