+++ app/api/backend/services/openai/generate-text/route.ts
