--- constants/api.urls.ts
@@ -1,7 +1,0 @@
-const ROUTES = {
-    START_JOB: "/api/start-job",
-    OPENAI: "/api/openai"
-  } as const;
-  
-  export default ROUTES;
-
