+++ constants/urls/services.urls.ts
@@ -1,0 +1,8 @@
+const TIKTRACK_SCRAPER_SYSTEM_URL = process.env.NEXT_PUBLIC_TIKTRACK_SCRAPER_SYSTEM_URL!;
+
+const ROUTES = {
+    TIKTRACK_SCRAPER_SYSTEM: TIKTRACK_SCRAPER_SYSTEM_URL,
+  } as const;
+
+export default ROUTES;
+
