+++ constants/urls/urls.ts
@@ -6,7 +6,6 @@ const ROUTES = {
   MESSAGE_EDIT: "/messages/[id]",
   SIGN_IN: "/sign-in",
   SIGN_UP: "/sign-up",
-  INFLUENCERS_MANAGEMENT: "http://127.0.0.1:8000/api/v1/",
   PROFILE: "/profile",
 } as const;
