+++ src/shared/utils/auth.util.ts
@@ -2,7 +2,7 @@
 
 import { cookies } from "next/headers";
 import { redirect } from "~/i18n/routing";
-import ROUTES from "~/constants/urls";
+import ROUTES from "~/constants/urls/urls";
 
 export async function logout(locale: string) {
   const cookieStore = await cookies();
